Add split save/load (de)serializers
Instead of writing a single serializer function which must be the same
for both serialization and deserialization, it is now possible to write
split save and load functions.

As part of the improvement, if no (de)serializer is found for a given
type, an explicit error message is now thrown.