Fix RPC calls with non-reference type parameters
The helper::functor_traits::foward_as_tuple function returns a lambda
function taking the Args&&... argument types (where Args... are the RPC
method argument types). Contrarily to what was expected, the perfect
forwarding template argument syntax only works for deduced arguments,
the compiler needs to know whether the variable is a lvalue or rvalue.

Since the lambda argument types are not deduced, the following happened:

* When a reference type was used for the RPC parameter, & collapsed with
  && into &, a lvalue reference.
* When a non-reference type was used for the RPC parameter, the
  "collapsed" was directly &&, a rvalue reference.

The solution consists in making sure that all values are captured by
contant lvalue reference in the lambda parameter. The compiler should be
able to optimize the binary as well as before the fix.